+++ src/utils/nodes/headerFooter/getAttributes.ts
